+++ tests/test_dummy_vpi/cfg.lua
@@ -1,0 +1,13 @@
+local cfg = {}
+
+cfg.simulator = _G.cfg.simulator
+if cfg.simulator == "verilator" then
+    cfg.top = "TOP.tb_top"
+else
+    cfg.top = "tb_top"
+end
+
+cfg.script = "./main.lua"
+cfg.is_hse = true
+
+return cfg